Move mask_grid function from utils to realtime_writer
