+++ handlers/realtime_writer.py
@@ -8,13 +8,13 @@ from typing import Optional
 
 from influxdb_client import Point, WritePrecision
 import numpy as np
-from shapely.geometry import shape
+from shapely.geometry import Point as GeoPoint, shape
 from shapely.prepared import prep, PreparedGeometry
 from xarray import Dataset
 
 from database.sql_manager import SqlManager
 from database.influx_manager import InfluxManager
-from procedures.utils.helpers import dt64_to_unixtime, mask_grid
+from procedures.utils.helpers import dt64_to_unixtime
 from handlers import config_handler
 from handlers.logging_handler import logger
 
@@ -227,6 +227,27 @@ class RealtimeWriter:
         thread = Thread(target=self.push_results, args=(rain_grids, x_grid, y_grid, calc_dataset))
         thread.start()
         return thread
+
+
+def mask_grid(
+        data_grid: np.ndarray,
+        x_grid: np.ndarray,
+        y_grid: np.ndarray,
+        prepared_polygons: list[PreparedGeometry]
+) -> np.ndarray:
+    """
+    Mask the 2D data grid with polygons. If a point is not within any of the polygons, it is set to NaN.
+
+    :param data_grid: 2D ndarray data grid to be masked
+    :param x_grid: 2D ndarray of x coordinates
+    :param y_grid: 2D ndarray of y coordinates
+    :param prepared_polygons: list of prepared polygons
+    :return: masked 2D ndarray with NaN values outside the polygons
+    """
+    mask = np.vectorize(lambda lon, lat: any(polygon.contains(GeoPoint(lon, lat)) for polygon in prepared_polygons))
+    within_polygons = mask(x_grid, y_grid)
+    data_grid[~within_polygons] = np.nan
+    return data_grid
 
 
 def _get_color(value: float) -> tuple[int, int, int, int]:
